chore: merged DatabaseInserter to DBInserter
