+++ application/src/main/java/org/utdteamthreefive/backend/service/DBInserter.java
@@ -1,7 +1,11 @@
 package org.utdteamthreefive.backend.service;
 
 import java.nio.file.Path;
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.SQLTransientException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -13,9 +17,10 @@ import java.util.concurrent.BlockingQueue;
 import java.util.logging.Logger;
 
 import org.utdteamthreefive.backend.models.Batch;
-import org.utdteamthreefive.backend.models.Batch.WordDelta;
 import org.utdteamthreefive.backend.models.Batch.BigramDelta;
+import org.utdteamthreefive.backend.models.Batch.WordDelta;
 import org.utdteamthreefive.backend.util.DatabaseManager;
+import org.utdteamthreefive.ui.Table;
 
 /**
  * DBInserter consumes parsed batches from a queue and inserts
@@ -24,115 +29,240 @@ import org.utdteamthreefive.backend.util.DatabaseManager;
  * Updated for the new global SentenceBuilder schema:
  * - WORD and WORD_FOLLOW are global (no per-file tables)
  * - FILES still logs file metadata (path, word count, date)
- * @Author Aisha Qureshi
+ * @author Aisha Qureshi, Zaeem Rashid, and Rommel Isaac Baldivas
  */
 public class DBInserter implements Runnable {
     private static final Logger logger = Logger.getLogger(DBInserter.class.getName());
-
-    private final Path path;
+    private Table table;
     private final BlockingQueue<Batch> batchQueue;
     private Connection conn;
-    private Long fileId;
+    private final Map<Path, Long> fileIds = new HashMap<>();
+    private final Map<Path, Long> wordsPerFile = new HashMap<>();
+    private int batchesProcessed = 0;
+    private int endMarkersProcessed = 0;
 
-    private long wordsThisFile = 0L;
-
-    public DBInserter(Path path, BlockingQueue<Batch> queue) {
-        this.path = path;
+    public DBInserter(BlockingQueue<Batch> queue, Table table) {
         this.batchQueue = queue;
-    }
-    private static boolean isTransient(SQLException e) {
-    int code = e.getErrorCode(); // MySQL vendor code
-        return code == 1213 || code == 1205 || e instanceof SQLTransientException;
+        this.table = table;
     }
 
-    private static long backoff(int attempt) {
-        return Math.min(2000L, (long) (50 * Math.pow(2, attempt)));
-    }
-
+    /**
+     * @author Aisha Qureshi and Rommel Isaac Baldivas
+     */
     @Override
     public void run() {
         try {
             conn = DatabaseManager.open();
-            if (conn == null) {
-                logger.severe("❌ Failed to open database connection");
-                return;
-            }
-
             conn.setAutoCommit(false);
-
-            // Lower isolation level to reduce deadlocks
             conn.setTransactionIsolation(Connection.TRANSACTION_READ_COMMITTED);
 
-            // Step 1: Ensure file record exists
-            fileId = ensureFileId();
-            if (fileId == null) {
-                logger.severe("❌ Could not create file record for " + path);
-                return;
-            }
+            // Continuously consume batches from the queue and insert/update DB
+            logger.info("DatabaseInserter started and waiting for batches...");
+            while (!Thread.currentThread().isInterrupted()) {
+                try {
+                    Batch batch = batchQueue.take();
 
-            logger.info("✅ DBInserter started for file: " + path);
-
-            boolean running = true;
-            while (running) {
-                Batch batch = batchQueue.take();
-
-                // Retry loop for transient errors
-                for (int attempt = 0;; attempt++) {
-                    try {
-                        insertOrUpdateWords(batch);
-                        insertOrUpdateBigrams(batch);
-                        if (batch.end) {
-                            updateFileTotals();
-                            running = false;
-                        }
-                        conn.commit();
-                        break; // ✅ success, exit retry loop
-
-                    } catch (SQLException e) {
-                        try {
-                            conn.rollback();
-                        } catch (SQLException ignore) {}
-
-                        if (isTransient(e) && attempt < 2) {
-                            long sleep = backoff(attempt);
-                            logger.warning("🔁 Deadlock/timeout, retrying batch in " + sleep + "ms: " + e.getMessage());
-                            Thread.sleep(sleep);
-                            continue;
-                        }
-
-                        logger.severe("❌ Fatal SQL after retries: " + e.getMessage());
-                        throw e;
+                    // Skip batches with null file path
+                    if (batch.filePath == null && !batch.end) {
+                        logger.warning("Skipping batch with null file path");
+                        continue;
                     }
+
+                    // Initialize file processing if not already done (for both regular and end batches)
+                    if (batch.filePath != null && !fileIds.containsKey(batch.filePath)) {
+                        Long fileId = ensureFileId(batch.filePath);
+                        if (fileId == null) {
+                            logger.severe("Could not create file record for " + batch.filePath.toString());
+                            continue; // Skip this batch but continue processing others
+                        }
+                        fileIds.put(batch.filePath, fileId);
+                        wordsPerFile.put(batch.filePath, 0L);
+                        logger.info("DatabaseInserter initialized for file: " + batch.filePath.toString());
+                    }
+
+                    // Handle regular batches
+                    if (batch.filePath != null && !batch.end) {
+                        batchesProcessed++;
+                        // Process the batch with retry logic
+                        boolean success = processBatchWithRetry(batch);
+                        if (!success) {
+                            logger.severe("Failed to process batch after all retries for file: " + batch.filePath.toString());
+                            // will continue processing other files even if current file fails
+                        }
+                    }
+
+                    // Handle end markers
+                    if (batch.end && batch.filePath != null) {
+                        endMarkersProcessed++;
+                        logger.info("🏁 Processing end marker for file: " + batch.filePath.getFileName() + " (end marker " + endMarkersProcessed + " of " + fileIds.size() + " files)");
+                        // Update file totals with retry logic for this specific file
+                        if (updateFileTotalsWithRetry(batch.filePath)) {
+                            logger.info("DatabaseInserter completed file: " + batch.filePath.toString() + " with " + wordsPerFile.get(batch.filePath) + " words");
+                        }
+                        else {
+                            logger.severe("Failed to update file totals after all retries for file: " + batch.filePath.toString());
+                        }
+
+                        logger.info("DatabaseInserter progress: " + endMarkersProcessed + "/" + fileIds.size() + " files completed");
+
+                        // Finally update the TableView
+                        table.syncTableWithDatabase();
+                    }
+                } catch (InterruptedException e) {
+                    // Interrupted - time to shutdown
+                    Thread.currentThread().interrupt();
+                    logger.info("DatabaseInserter was interrupted, shutting down...");
+                    break;
                 }
             }
 
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            logger.warning("⚠️ Inserter interrupted: " + e.getMessage());
-
-        } catch (SQLException e) {
-            logger.severe("💥 SQL error: " + e.getMessage());
-
+        } catch (Exception e) {
+            logger.severe("❌ DatabaseInserter failed: " + e.getMessage());
+            e.printStackTrace();
         } finally {
-            DatabaseManager.close(conn);
+            // Cleanup resources
+            if (conn != null) {
+                DatabaseManager.close(conn);
+                logger.info("DatabaseInserter completed processing " + fileIds.size() + " files" + " (processed " + batchesProcessed + " batches, " + endMarkersProcessed + " end markers)");
+            }
         }
+    }
+
+    ////#region HELPER METHODS
+    
+    /**
+     * Process a batch with proper retry logic for deadlock handling.
+     * @author Rommel Isaac Baldivas
+     */
+    private boolean processBatchWithRetry(Batch batch) {
+        final int MAX_RETRIES = 5; // Increased for deadlock scenarios
+        
+        for (int attempt = 0; attempt < MAX_RETRIES; attempt++) {
+            try {
+                insertOrUpdateWords(batch);
+                insertOrUpdateBigrams(batch);
+                conn.commit(); // Commit the transaction after successful batch processing
+                return true; // Success
+                
+            } catch (SQLException e) {
+                try {
+                    conn.rollback();
+                } catch (SQLException rollbackEx) {
+                    logger.severe("Failed to rollback transaction: " + rollbackEx.getMessage());
+                }
+
+                if (isTransient(e) && attempt < MAX_RETRIES - 1) {
+                    long sleep = backoff(attempt);
+                    logger.warning("Deadlock/timeout detected (attempt " + (attempt + 1) + "/" + MAX_RETRIES + 
+                                 "), retrying batch in " + sleep + "ms: " + e.getMessage());
+                    try {
+                        Thread.sleep(sleep);
+                    } catch (InterruptedException ie) {
+                        Thread.currentThread().interrupt();
+                        logger.severe("Thread interrupted during retry backoff");
+                        return false;
+                    }
+                    
+                    // Add jitter to reduce collision probability
+                    try {
+                        Thread.sleep((long)(Math.random() * 100));
+                    } catch (InterruptedException ie) {
+                        Thread.currentThread().interrupt();
+                        return false;
+                    }
+                    continue;
+                }
+
+                logger.severe("Fatal SQL error after " + MAX_RETRIES + " retries in " + 
+                             (batch.filePath != null ? batch.filePath.toString() : "unknown file") + ": " + e.getMessage());
+                e.printStackTrace();
+                return false;
+            } catch (Exception e) {
+                logger.severe("Unexpected error processing batch: " + e.getMessage());
+                e.printStackTrace();
+                return false;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Update file totals with retry logic for a specific file.
+     * @author Rommel Isaac Baldivas
+     */
+    private boolean updateFileTotalsWithRetry(Path filePath) {
+        final int MAX_RETRIES = 3;
+
+        for (int attempt = 0; attempt < MAX_RETRIES; attempt++) {
+            try {
+                updateFileTotals(filePath);
+                conn.commit(); // Final commit for file totals
+                return true;
+
+            } catch (SQLException e) {
+                try {
+                    conn.rollback();
+                } catch (SQLException rollbackEx) {
+                    logger.severe("Failed to rollback file totals transaction: " + rollbackEx.getMessage());
+                }
+
+                if (isTransient(e) && attempt < MAX_RETRIES - 1) {
+                    long sleep = backoff(attempt);
+                    logger.warning("Retrying file totals update in " + sleep + "ms: " + e.getMessage());
+                    try {
+                        Thread.sleep(sleep);
+                    } catch (InterruptedException ie) {
+                        Thread.currentThread().interrupt();
+                        return false;
+                    }
+                    continue;
+                }
+
+                logger.severe("Failed to update file totals after " + MAX_RETRIES + " retries: " + e.getMessage());
+                return false;
+            }
+        }
+        return false;
+    }
+    
+    /**
+     * 
+     * @param e
+     * @return
+     * @author Zaeem Rashid
+     */
+    private static boolean isTransient(SQLException e) {
+        int code = e.getErrorCode(); // MySQL vendor code
+        return code == 1213 || code == 1205 || e instanceof SQLTransientException;
+    }
+
+    /**
+     * 
+     * @param attempt
+     * @return
+     * @author Zaeem Rashid
+     */
+    private static long backoff(int attempt) {
+        return Math.min(2000L, (long) (50 * Math.pow(2, attempt)));
     }
     /**
      * Ensure file record exists, or insert a new one.
      */
-    private Long ensureFileId() throws SQLException {
+    private Long ensureFileId(Path filePath) throws SQLException {
         String selectSql = "SELECT file_id FROM files WHERE file_path = ?";
         String insertSql = "INSERT INTO files (file_path, word_count, date_imported) VALUES (?, 0, CURDATE())";
 
+        // Query for existing record
         try (PreparedStatement selectStmt = conn.prepareStatement(selectSql)) {
-            selectStmt.setString(1, path.toString());
+            selectStmt.setString(1, filePath.toString());
             try (ResultSet rs = selectStmt.executeQuery()) {
                 if (rs.next()) return rs.getLong("file_id");
             }
         }
 
+        // If it doesn't exist, insert new record
         try (PreparedStatement insertStmt = conn.prepareStatement(insertSql, PreparedStatement.RETURN_GENERATED_KEYS)) {
-            insertStmt.setString(1, path.toString());
+            insertStmt.setString(1, filePath.toString());
             insertStmt.executeUpdate();
             try (ResultSet rs = insertStmt.getGeneratedKeys()) {
                 if (rs.next()) return rs.getLong(1);
@@ -142,31 +272,38 @@ public class DBInserter implements Runnable {
     }
 
     /**
-         * Insert or update global WORD table.
-         */
-        private void insertOrUpdateWords(Batch batch) throws SQLException {
+     * Insert or update global WORD table.
+     */
+    private void insertOrUpdateWords(Batch batch) throws SQLException {
         if (batch.wordDeltas.isEmpty()) return;
 
         String upsertSql = """
-            INSERT INTO word (word_token, total_count, start_count, end_count, type_)
-            VALUES (?, ?, ?, ?, ?)
-            ON DUPLICATE KEY UPDATE
-                total_count = total_count + VALUES(total_count),
-                start_count = start_count + VALUES(start_count),
-                end_count = end_count + VALUES(end_count),
-                type_ = VALUES(type_)
-            """;
-        
+                INSERT INTO word (word_token, total_count, start_count, end_count, type_)
+                VALUES (?, ?, ?, ?, ?)
+                ON DUPLICATE KEY UPDATE
+                    total_count = total_count + VALUES(total_count),
+                    start_count = start_count + VALUES(start_count),
+                    end_count = end_count + VALUES(end_count),
+                    type_ = VALUES(type_)
+                """;
+
         long batchTotal = 0L;
 
         try (PreparedStatement stmt = conn.prepareStatement(upsertSql)) {
-            for (WordDelta wd : batch.wordDeltas) {
+            // Sort word deltas by token to reduce deadlock probability
+            List<WordDelta> sortedDeltas = new ArrayList<>(batch.wordDeltas);
+            sortedDeltas.sort((a, b) -> a.token().compareTo(b.token()));
+
+            int batchCount = 0;
+            final int BATCH_SIZE = 500; // Smaller batches to reduce deadlock window
+
+            for (WordDelta wd : sortedDeltas) {
                 stmt.setString(1, wd.token());
                 stmt.setInt(2, wd.total());
                 stmt.setInt(3, wd.begin());
                 stmt.setInt(4, wd.end());
 
-                // 🧠 Normalize type to match ENUM('alpha', 'misc')
+                // Normalize type to match ENUM('alpha', 'misc')
                 String type = wd.type().toLowerCase();
                 if (!type.equals("alpha")) {
                     type = "misc";
@@ -174,25 +311,39 @@ public class DBInserter implements Runnable {
                 stmt.setString(5, type);
 
                 batchTotal += wd.total();
-
                 stmt.addBatch();
-            }
-            stmt.executeBatch();
-        }
-        wordsThisFile += batchTotal;
-    }
+                batchCount++;
 
+                // Execute in smaller chunks to reduce deadlock window
+                if (batchCount >= BATCH_SIZE) {
+                    stmt.executeBatch();
+                    batchCount = 0;
+                }
+            }
+
+            // Execute remaining statements
+            if (batchCount > 0) {
+                stmt.executeBatch();
+            }
+        }
+
+        // Update word count for this specific file
+        if (batch.filePath != null) {
+            wordsPerFile.merge(batch.filePath, batchTotal, Long::sum);
+        }
+    }
+    
     private Map<String, Long> fetchWordIdsBulk(Connection conn, Set<String> tokens) throws SQLException {
         Map<String, Long> out = new HashMap<>(tokens.size() * 2);
         if (tokens.isEmpty()) return out;
-
+        
         // MySQL IN list practical chunking (avoid huge packets)
         final int CHUNK = 1000;
         List<String> list = new ArrayList<>(tokens);
         for (int i = 0; i < list.size(); i += CHUNK) {
             int end = Math.min(i + CHUNK, list.size());
             List<String> sub = list.subList(i, end);
-
+            
             String placeholders = String.join(",", Collections.nCopies(sub.size(), "?"));
             String sql = "SELECT word_id, word_token FROM word WHERE word_token IN (" + placeholders + ")";
             try (PreparedStatement ps = conn.prepareStatement(sql)) {
@@ -207,9 +358,11 @@ public class DBInserter implements Runnable {
         }
         return out;
     }
-
+    
+    
     /**
      * Insert or update global WORD_FOLLOW table.
+     * @author Aisha Qureshi and Rommel Isaac Baldivas
      */
     private void insertOrUpdateBigrams(Batch batch) throws SQLException {
         if (batch.bigramDeltas.isEmpty()) return;
@@ -224,42 +377,85 @@ public class DBInserter implements Runnable {
         // 2) Resolve IDs in bulk (words should already exist from insertOrUpdateWords in this same transaction)
         Map<String, Long> idMap = fetchWordIdsBulk(conn, needed);
 
-        // 3) Batch insert/merge
+        // 3) Sort bigrams by from_word_id, then to_word_id to reduce deadlock probability
+        List<BigramDelta> sortedBigrams = new ArrayList<>(batch.bigramDeltas);
+        sortedBigrams.sort((a, b) -> {
+            Long aFromId = idMap.get(a.key().first());
+            Long bFromId = idMap.get(b.key().first());
+            if (aFromId == null && bFromId == null) return 0;
+            if (aFromId == null) return 1;
+            if (bFromId == null) return -1;
+
+            int fromCompare = aFromId.compareTo(bFromId);
+            if (fromCompare != 0) return fromCompare;
+
+            Long aToId = idMap.get(a.key().second());
+            Long bToId = idMap.get(b.key().second());
+            if (aToId == null && bToId == null) return 0;
+            if (aToId == null) return 1;
+            if (bToId == null) return -1;
+
+            return aToId.compareTo(bToId);
+        });
+
+        // 4) Batch insert/merge with smaller batch sizes
         String upsert = """
-            INSERT INTO word_follow (from_word_id, to_word_id, total_count)
-            VALUES (?, ?, ?)
-            ON DUPLICATE KEY UPDATE total_count = total_count + VALUES(total_count)
-            """;
+                INSERT INTO word_follow (from_word_id, to_word_id, total_count)
+                VALUES (?, ?, ?)
+                ON DUPLICATE KEY UPDATE total_count = total_count + VALUES(total_count)
+                """;
 
         try (PreparedStatement ps = conn.prepareStatement(upsert)) {
             int batchSize = 0;
-            for (BigramDelta bd : batch.bigramDeltas) {
+            final int MAX_BATCH_SIZE = 500; // Smaller batches to reduce deadlock window
+
+            for (BigramDelta bd : sortedBigrams) {
                 Long fromId = idMap.get(bd.key().first());
-                Long toId   = idMap.get(bd.key().second());
+                Long toId = idMap.get(bd.key().second());
                 if (fromId == null || toId == null) continue; // should be rare if words inserted first
 
                 ps.setLong(1, fromId);
                 ps.setLong(2, toId);
                 ps.setInt(3, bd.count());
                 ps.addBatch();
+                batchSize++;
 
-                if (++batchSize % 1000 == 0) ps.executeBatch(); // reasonable chunk
+                // Execute in smaller chunks to reduce deadlock window
+                if (batchSize >= MAX_BATCH_SIZE) {
+                    ps.executeBatch();
+                    batchSize = 0;
+                }
             }
-            ps.executeBatch();
+
+            // Execute remaining statements
+            if (batchSize > 0) {
+                ps.executeBatch();
+            }
         }
     }
 
+    
     /**
-     * After all batches are processed, compute total words for the file.
-     */
-    private void updateFileTotals() throws SQLException {
+     * After all batches are processed, compute total words for the specific file.
+     * @author Aisha Qureshi and Rommel Isaac Baldivas
+    */
+   private void updateFileTotals(Path filePath) throws SQLException {
+       Long fileId = fileIds.get(filePath);
+       Long wordCount = wordsPerFile.get(filePath);
+       
+       if (fileId == null || wordCount == null) {
+           logger.warning("Cannot update totals for file: " + filePath + " - missing data");
+           return;
+        }
+        
         String updateSql = "UPDATE files SET word_count = ? WHERE file_id = ?";
         try (PreparedStatement stmt = conn.prepareStatement(updateSql)) {
-            stmt.setLong(1, wordsThisFile);
+            stmt.setLong(1, wordCount);
             stmt.setLong(2, fileId);
             stmt.executeUpdate();
         }
-
-        logger.info("📊 Updated totals → Words (this file): " + wordsThisFile);
+        
+        logger.info("Updated totals for " + filePath.getFileName() + " → Words: " + wordCount);
     }
+    ////#endregion
 }
